Move WriterScalar to separate file
