+++ src/CMakeLists.txt
@@ -158,6 +158,7 @@ set(kafka_to_nexus_INC
   schemas/ev42/ev42_synth.h
   schemas/f142/f142_rw.h
   schemas/f142/WriterTypedBase.h
+  schemas/f142/WriterScalar.h
   schemas/f142/WriterArray.h
   schemas/f142/f142_synth.h
   schemas/template/TemplateWriter.h
