+++ src/schemas/f142/f142_rw.h
@@ -11,19 +11,6 @@ namespace FileWriter {
 namespace Schemas {
 namespace f142 {
 
-template <typename DT, typename FV>
-class WriterScalar : public WriterTypedBase {
-public:
-  WriterScalar(hdf5::node::Group hdf_group, std::string const &source_name,
-               Value fb_value_type_id, CollectiveQueue *cq);
-  WriterScalar(hdf5::node::Group hdf_group, std::string const &source_name,
-               Value fb_value_type_id, CollectiveQueue *cq,
-               HDFIDStore *hdf_store);
-  h5::append_ret write_impl(FBUF const *fbuf) override;
-  uptr<h5::h5d_chunked_1d<DT>> ds;
-  Value _fb_value_type_id = Value::NONE;
-};
-
 class FlatbufferReader : public FileWriter::FlatbufferReader {
   bool verify(Msg const &msg) const override;
   std::string source_name(Msg const &msg) const override;
